Potentially here a major bug : number of schools per cohort at initialization is bigger than in the reproduce function. It leads to an artificial loss of biomass. Still to be thoroughly investigated.
(School.java)
Set abundance as double in the constructor.

(Cohort.java)
Deleted the growth function, as it was an unnecessary redundant in between species and school. It is now all in species.
In constructor, made sure the number of initial school is the same than the reproduce function in species. And no more surplus to redistribute since abundance is double precision.

(Species.java)
Created function initNumberSchools() that compute the number of schools per cohort depending on longevity and parameter "number of school per annual cohort".
Uses this function in both reproduce() and iniByBiomass()
In reproduce() no more surplus since abundance is double precision.

(Simulation.java)
Created function updateBiomassAndAbundance(). Called it after computeMortality(). Deleted unnecessary functions in newStep() such as assesAbdCht, assessVecPreSchool, etc.

git-svn-id: https://subversion.renater.fr/authscm/barrier.n/svn/osmose/trunk@249 45b28200-7c74-4d1a-96bc-342942c5875a