This commit changed the way of moving schools with connectivity matrices. It is now quite similar to map distribution. The difference comes at redistributing the schools from one map to another. In map distribution mode it is done randomly. In connectivity mode it is done through the connectivity matrix.
(Osmose.java)
Changed readConnectivity function. Every map of distribution comes with a matrix of connectivity, for connectivity mode.

(Simulation.java)
Modified connectivityDistribution() so that it distributes schools in map of distribution and when map changes, the application uses the matrix of connectivity to redistribute the school from one map to the next one.

git-svn-id: https://subversion.renater.fr/authscm/barrier.n/svn/osmose/trunk@352 45b28200-7c74-4d1a-96bc-342942c5875a