This is a major commit: all schools are now stored in a Population object.
Introduced FilteredSet.java, IFilter.java and some tools in FilteredSets.java. It is basically how to get subset of any list according to one or several filters.
Included a few filters, like SpeciesFilter, 

(Species.java)
Does not extend an arraylist of Schools.
Moved reproduce and incomingFlux method at Simulation level.

Deleted Cohort.java as it is not used any more. Could be replaced by a FilteredSet easily if needed.

(Simulation.java)
Added a population object that stores all schools and modified the methods that create new schools to add them into the population instead of Species.

At that stage, we commit all this new classes but it should be thouroughly checked further on and it is not optimized at all.

git-svn-id: https://subversion.renater.fr/authscm/barrier.n/svn/osmose/trunk@378 45b28200-7c74-4d1a-96bc-342942c5875a