Doing a bit of profiling I noticed that the function Simulation.getAccessibleCells was taking a lot of CPU time. I worked out a way to find accessible cells in a map avoiding the time consuming list.contains(cell) method. Created class GridMap that is a float[][] same size than the grid and contains the value of the probability of presence. An accessible cell in GridMap is a cell which value is greater than zero.
(Osmose.java)
Redefined maps as an array of GridMap. Modified readCSVMap to load the value in the gridMaps.

(Simulation.java)
Modified mapDistribution and connectivityDistribution to handle the new GridMap object. And changed the getAccessibleCells function so that it only has to check whether neighbours cells have values > 0. It seems much faster.

(Cell.java)
Added a constant for value of LAND (-99). It must be a negative value.

git-svn-id: https://subversion.renater.fr/authscm/barrier.n/svn/osmose/trunk@379 45b28200-7c74-4d1a-96bc-342942c5875a