The same way I created some fishing scenario classes, I added a new package process.naturalmortality to include different types of natural mortality parametrizations: constant D, constant larva mortality, ByDtLarvaMortality, ByDtByAgeSize natural mortality, etc.
Transformed AbstractFishingScenario.jva into AbstractMortalityScenario.java that can be used by both fishing and natural mortality.

Created class ByClassTimeSeries.java that is able to read a CSV file with values in a matrix with lines as time steps an columns as classes. The class read as well the value of the thresholds in the header of the first line.

git-svn-id: https://subversion.renater.fr/authscm/barrier.n/svn/osmose/trunk@808 45b28200-7c74-4d1a-96bc-342942c5875a