Big commit for mortality algorithm and fishing mortality by catches.
Deleted mortality algorithm CASE2 (INDEPENDENT).
Updated STOCHASTIC case so that it now works with both F and catches. I had to change quite a lot of things for not loosing time at estimating the fishable biomass for every species at every subdt. I had to swap the loops on cells and subdt, to cache the list of schools per species in SchoolSet.java.

Added abstract function assessFishableBiomass() in AbstractFishingScenario.java and implemented it for all the catches scenario.

git-svn-id: https://subversion.renater.fr/authscm/barrier.n/svn/osmose/trunk@930 45b28200-7c74-4d1a-96bc-342942c5875a