Renamed RestartPopulator.java into NetcdfPopulator.java
Added a new way of initializing population: 'population.initialization.file' which similar to the NetCDF restart file. The only difference is that the simulation starts at step 0, the global attribute from the restart file is not read.
Added an entry for the parameter 'population.initialization.method' that can also take 'netcdf' from now.

git-svn-id: https://subversion.renater.fr/authscm/barrier.n/svn/osmose/trunk@969 45b28200-7c74-4d1a-96bc-342942c5875a