Commit r1020 is a mess and gives compilation error. At reverting r1019, I forgot to commit Prey.java. And for r1020 I forgot also to commit BiomassPopulator.java This new commit replaces r1020.
The same way we clarified the meaning of the accessibility coefficient at r1018, by saying that plankton biomass * accessibility coefficient (the one estimated by the calibration) is the biomass of plankton available to schools for predation process; we also clarifies the meaning of the larval mortality. We now say that the larval mortality is actually an egg loss prior to any predation process. The egg mortality applies to the school before entering the subdt loop of predation. Therefore abundance of egg at the beginning of the predation process is abundance*exp(-D) with D the egg mortality estimated by the calibration. This modification also implies that egg won't suffer any additional natural mortality in the mortality subdt loop (other way to put it is to say that the only mortality that will affect eggs in the subdt loop is predation as fishing and starvation mortality were already null).
Then we still want to try whether the remaining eggs should all be released at the beginning of the subdt loop or, on the contrary, progressively released over the subdt as it was done at r1019. To try this, we added a temporary boolean in MortalityProcess.java, eggRetained = true/false.
EggRetained = false means that all the eggs left after initial natural mortality are available for predation from the 1st subdt. 
EggRetained = true means that the eggs left after initial natural mortality are released progressively over the sub time step. At the beginning of every sub time step instantaneous abundance of eggs is incremented by (initial abundance - eggLoss)/subdt


git-svn-id: https://subversion.renater.fr/authscm/barrier.n/svn/osmose/trunk@1021 45b28200-7c74-4d1a-96bc-342942c5875a