(ReproductionProcess.java) Added a new important feature to Osmose that might replace the current initialisation from target biomass : population seeding. The user sets, for every species, a seeding biomass, which is seen by Osmose as the Spawning Stock Biomass to take into consideration when the reproductive individuals are extinct. In practical terms it means two things: 1. the model will be initialised by a flux of eggs as long as the first released eggs reach maturity. 2. if a species collapses, or is left without any reproductive individuals (SSB == 0), Osmose will use the seeding biomass as backup SSB to remedy the collapse Parameters 'population.seeding.biomass.sp#' or 'population.seeding.biomass'. All these values are set to zero by default, meaning no seeding.
The user must also set a year max for the seeding process. Parameter 'population.seeding.year.max'. If the parameter does not exist, the year max is set to the lifespan of the species that live the longest. 

Concluding remark: I did not unplug 'population.initialization.*' parameters, since the seeding is still an experimental feature. When using the seeding, the user must make sure that the usual initialisation biomass  'population.initialisation.biomass.sp#' are set to zero. Otherwise population initialisation and population seeding processes will overlap.
One should also be careful with the species modelled with an input flux of biomass. In such case, the seeding biomass should be set to zero to avoid any interference.

git-svn-id: https://subversion.renater.fr/authscm/barrier.n/svn/osmose/trunk@1084 45b28200-7c74-4d1a-96bc-342942c5875a