(Osmose.java) Added new optional command line option -resolve=global|local -resolve=global means that every relative path in Osmose configuration file will be resolved against the path of the main configuration file. -resolve=local means that every relative path in Osmose configuration file will be resolved against the path of the current configuration file (the one that actually contains the path parameter).
(Configuration.java)
Added a new Properties variable cmd that receives all the command line options (only one for now). In the init() function the command line options are added to the list of Osmose parameters (cfg variable). As a consequence an option that has been defined in the command line can overwrite a parameter from the Osmose configuration file.

git-svn-id: https://subversion.renater.fr/authscm/barrier.n/svn/osmose/trunk@1306 45b28200-7c74-4d1a-96bc-342942c5875a