(AdditionalMortality.java) Used enum to list every Larva Additional Mortality scenarios. Make sure that the configuration file exactly contains one Additional Mortality scenario, and throw error otherwise
(ConstantLarvaMortality.java)
Deleted default larval mortality rate. Osmose used to throw a warning when no larva mortality rate was defined and set it to zero, taking into account that for some species that do not reproduce in situ, this mortality may not apply. I'd rather throw an error and force the user to manually set it to zero so Osmose does not make any assumption.

git-svn-id: https://subversion.renater.fr/authscm/barrier.n/svn/osmose/trunk@1394 45b28200-7c74-4d1a-96bc-342942c5875a