This commit changes how the LTL is handled in Osmose and intents to be less error prone. Most importantly the commit updates the meaning of the conversion factor (plankton.conversion2tons.plk#). It used to be the conversion from LTL unit to Osmose unit (tonne of wet weight) after the vertical integration and it included conversion from square meter to square kilometre, which was confusing. From now on, conversion factor plankton.conversion2tons.plk# has the following meaning : it converts from LTL unit per cubic metre into tonne of wet weight per cubic metre. Osmose will first convert the LTL density into tonne of wet weight and afterwards integrates vertically and per cell.
(AbstractLTLForcing.java) 
getRawBiomass() still exists and is still abstracts. The return type is now a 3D array, without vertical integration. The vertical integration is handled in the update() function, before vertical integration (new function zIntegration()) and spatial integration.

(AbstractLTLFastForcing.java)
I also changed the way I handled "fast" forcing. Fast forcing only means that all LTL data is loaded in the heap at the beginning of the simulation. So actually the fastforcing is going to be the same no matter whether we read PISCES, BFM or ECO3M. AbstractLTLFastForcing only needs to know what is the class that handles the LTL and to call it for loading at first time step all the LTL data. I kept for the moment the LTLFastForcingECO3M/RomsPisces/BFM.java classes but they will be deleted in near future as they have become empty shells that only link AbstractLTLFastForcing to the correct LTL plug-in.

(LTLFastForcing.java)
So far the LTLFastForcing does not change much and still does not take into account the conversion factor. We are going to change this in future commits.

(LTLForcing.java)
(re)Added interface LTLForcing.java since we needed a parent class for both AbstractLTLForcing.java and AbstractLTLFastForcing.java.

Deleted LTLUtil.java as the verticalIntegration() function has been delegated to AbstractLTLForcing.java

git-svn-id: https://subversion.renater.fr/authscm/barrier.n/svn/osmose/trunk@1403 45b28200-7c74-4d1a-96bc-342942c5875a